fix: route structure of bookings index
