+++ app/routes/_layout+/kits.$kitId.bookings.tsx
@@ -20,7 +20,7 @@ import {
   PermissionEntity,
 } from "~/utils/permissions/permission.data";
 import { requirePermission } from "~/utils/roles.server";
-import BookingsIndexPage from "./bookings";
+import BookingsIndexPage from "./bookings._index";
 
 const BOOKING_STATUS_TO_SHOW = [
   BookingStatus.DRAFT,
